userInput.java
About BufferReader and FileReader.

The BufferedReader can be used to read a file.
FileReader is a class in Java used to read data from a file in a character stream format.
FileReader is used to read character files (text files).

The BufferedReader is used to efficiently read the content of the file. The readLine() method reads the file line by line, which is useful for processing text files.